Modification of delete routes in sauce controlers

+++ controllers/sauce.js
@@ -78,31 +78,50 @@ exports.updateSauce = ((req, res, next) => {
 
 
 exports.deleteSauce = ((req, res, next) => {
-    Sauce.deleteOne({
-            _id: req.params.id
-        })
-        .then(() => res.status(200).json({
-            message: 'Objet supprimé !'
-        }))
-        .catch(error => res.status(400).json({
-            error
-        }));
-});
+            Sauce.findOne({
+                    _id: req.params.id
+                })
+                .then(sauce => {
+                        if (sauce.userId != req.auth.userId) {
+                            res.status(401).json({
+                                message: 'non-autorisé'
+                            })
+                        } else {
+                            const filename = sauce.imageUrl.split('/images/')[1];
+                            fs.unlink(`images/${filename}`, () => {
+                                Sauce.deleteOne({
+                                        _id: req.params.id
+                                    })
+                                    .then(() => {
+                                        res.status(200).json({
+                                            message:"Object supprimé"
+                                        })
+                                    })
+                                    .catch(error => res.status(401).json({error}));
+                            })
 
-exports.getOneSauce = ((req, res, next) => {
-    Sauce.findOne({
-            _id: req.params.id
-        })
-        .then(sauce => res.status(200).json(sauce))
-        .catch(error => res.status(400).json({
-            error
-        }));
-});
+                        }
+                    })
 
-exports.getAllSauce = ((req, res, next) => {
-    Sauce.find()
-        .then(sauce => res.status(200).json(sauce))
-        .catch(error => res.status(400).json({
-            error
-        }));
-});
+                .catch(error => res.status(500).json({
+                    error
+                }));
+            });
+
+        exports.getOneSauce = ((req, res, next) => {
+            Sauce.findOne({
+                    _id: req.params.id
+                })
+                .then(sauce => res.status(200).json(sauce))
+                .catch(error => res.status(400).json({
+                    error
+                }));
+        });
+
+        exports.getAllSauce = ((req, res, next) => {
+            Sauce.find()
+                .then(sauce => res.status(200).json(sauce))
+                .catch(error => res.status(400).json({
+                    error
+                }));
+        });
